Refactoring documents and DocumentLists
