+++ client/app/reader/Documents/actionTypes.js
@@ -1,0 +1,16 @@
+/** Document Related */
+export const ASSIGN_DOCUMENTS = 'ASSIGN_DOCUMENTS';
+export const RECEIVE_DOCUMENTS = 'RECEIVE_DOCUMENTS';
+export const ROTATE_PDF_DOCUMENT = 'ROTATE_PDF_DOCUMENT';
+export const SELECT_CURRENT_VIEWER_PDF = 'SELECT_CURRENT_VIEWER_PDF';
+
+// Tags
+export const REQUEST_NEW_TAG_CREATION_SUCCESS = 'REQUEST_NEW_TAG_CREATION_SUCCESS';
+export const REQUEST_NEW_TAG_CREATION_FAILURE = 'REQUEST_NEW_TAG_CREATION_FAILURE';
+export const REQUEST_NEW_TAG_CREATION = 'REQUEST_NEW_TAG_CREATION';
+export const REQUEST_REMOVE_TAG = 'REQUEST_REMOVE_TAG';
+export const REQUEST_REMOVE_TAG_SUCCESS = 'REQUEST_REMOVE_TAG_SUCCESS';
+export const REQUEST_REMOVE_TAG_FAILURE = 'REQUEST_REMOVE_TAG_FAILURE';
+
+export const ROTATION_INCREMENTS = 90;
+export const COMPLETE_ROTATION = 360;
