+++ client/test/node/reader/DecisionReviewer-test.js
@@ -18,7 +18,8 @@ import { formatDateStr } from '../../../app/util/DateUtil';
 
 import readerReducer from '../../../app/reader/reducer';
 import caseSelectReducer from '../../../app/reader/CaseSelect/CaseSelectReducer';
-import { annotationLayerReducer } from '../../../app/reader/AnnotationLayer/AnnotationLayerReducer';
+import annotationLayerReducer from '../../../app/reader/AnnotationLayer/AnnotationLayerReducer';
+import documentListReducer from '../../../app/reader/DocumentList/DocumentListReducer';
 
 import PdfJsStub, { PAGE_WIDTH, PAGE_HEIGHT } from '../../helpers/PdfJsStub';
 import { onReceiveDocs } from '../../../app/reader/LoadingScreen/LoadingScreenActions';
@@ -60,6 +61,7 @@ describe('DecisionReviewer', () => {
         caseSelect: caseSelectReducer,
         readerReducer,
         search: searchReducer,
+        documentList: documentListReducer,
         annotationLayer: annotationLayerReducer
       }),
       compose(
